Tools for sending moves.
The server can now be configured with optional gas settings and
a private key for an operator address.  We also added tools
for checking whether this operator address or one configured on the
fly can send a particular move (either directly or via delegation),
and one to actually send a move in this way.

It also has a tool to check for the status (and optionally wait for
confirmation) of one or multiple transactions.